requirements.txt
decorator==4.0.4
Django==1.8.5
djangorestframework==3.2.4
docutils==0.12
requests==2.8.1
robotframework==2.9.2
robotframework-requests==0.3.8
robotframework-selenium2library==1.7.4
rest-framework-generic-relations==1.1.0
selenium==2.48.0
